Configure background fetch during app widget init
